fix comment

+++ src/main/java/com/example/electronicshop/controller/CommentController.java
@@ -18,46 +18,56 @@ import javax.validation.Valid;
 
 @RestController
 @AllArgsConstructor
-@RequestMapping("/api/comment")
+@RequestMapping("/api")
 public class CommentController {
     private final CommentService commentService;
     private final JwtUtils jwtUtils;
 
     @GetMapping(path = "/comment/{productId}")
-    public ResponseEntity<?> findByProductId (@PathVariable("productId") String productId,
-                                              @PageableDefault(size = 5, sort = "createdDate, DESC") @ParameterObject Pageable pageable){
+    public ResponseEntity<?> findByProductId(@PathVariable("productId") String productId,
+                                             @PageableDefault(size = 5, sort = "createdDate, DESC") @ParameterObject Pageable pageable) {
         return commentService.findByProductId(productId, pageable);
     }
 
     @GetMapping(path = "/admin/manage/comment/findall")
-    public ResponseEntity<?> findAllComment (){
+    public ResponseEntity<?> findAllComment() {
         return commentService.findAllComment();
     }
 
 
     @PostMapping(path = "/comment")
-    public ResponseEntity<?> addComment (@Valid @RequestBody CommentReq req,
-                                         HttpServletRequest request){
+    public ResponseEntity<?> addComment(@Valid @RequestBody CommentReq req,
+                                        HttpServletRequest request) {
         User user = jwtUtils.getUserFromJWT(jwtUtils.getJwtFromHeader(request));
         if (!user.getId().isBlank())
             return commentService.addComment(user.getId(), req);
         throw new AppException(HttpStatus.FORBIDDEN.value(), "You don't have permission! Token is invalid");
     }
 
-    @PutMapping("/admin/manage/comment/setenable/{CommentId}")
-    public ResponseEntity<?> setEnableComment (@PathVariable("CommentId") String CommentId)
-    {
-        return commentService.setEnableComment(CommentId);
-    }
-    @DeleteMapping("/admin/manage/comment/block/{CommentId}")
-    public ResponseEntity<?> blockComment (@PathVariable("CommentId") String CommentId)
-    {
-        return commentService.blockComment(CommentId);
-    }
-    @DeleteMapping("/admin/manage/comment/delete/{CommentId}")
-    public ResponseEntity<?> deleteComment (@PathVariable("CommentId") String CommentId)
-    {
-        return commentService.deleteCommemt(CommentId);
+    @PutMapping(path = "/comment/edit/{CommentId}")
+    public ResponseEntity<?> editCommentbyUser(@PathVariable("CommentId") String CommentId
+            , @RequestBody CommentReq req, HttpServletRequest request) {
+        User user = jwtUtils.getUserFromJWT(jwtUtils.getJwtFromHeader(request));
+        if (!user.getId().isBlank())
+            return commentService.editCommentByUser(CommentId, user.getId(), req);
+        throw new AppException(HttpStatus.FORBIDDEN.value(), "You don't have permission! Token is invalid");
     }
 
+    @PutMapping("/admin/manage/comment/setenable/{CommentId}")
+    public ResponseEntity<?> setEnableComment(@PathVariable("CommentId") String CommentId) {
+        return commentService.setEnableComment(CommentId);
+    }
+
+    @DeleteMapping("/admin/manage/comment/block/{CommentId}")
+    public ResponseEntity<?> blockComment(@PathVariable("CommentId") String CommentId) {
+        return commentService.blockComment(CommentId);
+    }
+
+    @DeleteMapping("/comment/delete/{CommentId}")
+    public ResponseEntity<?> deleteComment(@PathVariable("CommentId") String CommentId, HttpServletRequest request) {
+        User user = jwtUtils.getUserFromJWT(jwtUtils.getJwtFromHeader(request));
+        if (!user.getId().isBlank())
+            return commentService.deleteCommemt(CommentId);
+        throw new AppException(HttpStatus.FORBIDDEN.value(), "You don't have permission! Token is invalid");
+    }
 }
